dep(frontend): npm updates refactor(frontend)
