+++ frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "faucet-app",
-  "version": "0.4.0",
+  "version": "0.5.0",
   "private": true,
   "scripts": {
     "start": "react-scripts start",
@@ -11,32 +11,29 @@
     "build-storybook": "build-storybook -s public"
   },
   "dependencies": {
-    "axios": "^0.19",
+    "axios": "^0.21.1",
     "bech32": "^1.1.3",
     "classnames": "^2.2.6",
     "formik": "^2",
-    "husky": "^4",
-    "lint-staged": "^10",
-    "node-sass": "^4.11.0",
-    "prettier": "^2",
-    "react": "^16",
-    "react-dom": "^16",
-    "react-router-dom": "^5.0.0",
-    "react-scripts": "^3",
-    "react-toastify": "^6",
-    "reaptcha": "^1.4.2",
+    "react": "^17",
+    "react-dom": "^17",
+    "react-google-recaptcha": "^2.1.0",
+    "react-router-dom": "^5",
+    "react-scripts": "^4",
+    "react-toastify": "^7",
+    "sass": "^1.32.12",
     "typescript": "^4",
-    "yup": "^0.29"
+    "yup": "^0.32"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.0",
-    "@storybook/addon-actions": "^6",
-    "@storybook/addon-links": "^6",
-    "@storybook/addons": "^6",
-    "@storybook/react": "^6",
-    "babel-loader": "8.1.0",
+    "@babel/core": "^7",
+    "babel-loader": "^8",
+    "husky": "^4",
+    "lint-staged": "^10",
+    "prettier": "^2",
     "regenerator-runtime": "^0.13",
-    "webpack": "4.42.0"
+    "webpack": "4.44.2",
+    "webpack-cli": "^4.2.0"
   },
   "lint-staged": {
     "src/**/*.{js,jsx,ts,tsx,json,css,scss,md}": [
