+++ frontend/src/App.scss
@@ -26,7 +26,8 @@
 }
 
 header {
-  background: #2845ae url(/assets/terra_faucet.svg) top 15px left 20px no-repeat;
+  background: #2845ae url(./assets/terra_faucet.svg) top 15px left 20px
+    no-repeat;
   background-size: 120px auto;
   text-align: right;
   color: #fff;
