+++ frontend/src/config/networks.js
@@ -1,6 +1,8 @@
-export default [
+const networks = [
   {
     key: 'tequila-0004',
     lcd: 'https://tequila-fcd.terra.dev',
   },
 ];
+
+export default networks;
