--- frontend/src/route.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import { Route } from 'react-router-dom';
-import Home from './Pages/Home';
-
-export default function () {
-  return (
-    <React.Fragment>
-      <Route exact path="/" component={Home} />
-    </React.Fragment>
-  );
-}
